Get webhook IP/Domain in github_webhook_server container

+++ github-webhook-server-container/webhook.py
@@ -15,27 +15,31 @@ def _get_firefox_driver():
 
 
 def create_webhook():
+    ngrok_url = None
     github_owner = os.environ["GITHUB_OWNER"]
     github_repository = os.environ["GITHUB_REPOSITORY"]
     github_token = os.getenv("GITHUB_TOKEN")
     events = os.environ["GITHUB_EVENTS"].split(",")
-    driver = _get_firefox_driver()
-    driver.get("http://ngrok:4040/status")
-    ngrok_url = driver.find_element(
-        "xpath",
-        '//*[@id="app"]/div/div/div/div[1]/div[1]/ul/li[1]/div/table/tbody/tr[1]/td',
-    ).text
-    driver.close()
-    config = {"url": f"{ngrok_url}/github_webhook", "content_type": "json"}
-    gapi = Github(login_or_token=github_token)
-    repo = gapi.get_repo(f"{github_owner}/{github_repository}")
-    for _hook in repo.get_hooks():
-        if "ngrok.io" in _hook.config["url"]:
-            print(f"Deleting existing webhook: {_hook.config['url']}")
-            _hook.delete()
+    webhook_ip = os.environ.get("GITHUB_WEBHOOK_IP")
+
+    if webhook_ip == "ngrok":
+        driver = _get_firefox_driver()
+        driver.get("http://ngrok:4040/status")
+        ngrok_url = driver.find_element(
+            "xpath",
+            '//*[@id="app"]/div/div/div/div[1]/div[1]/ul/li[1]/div/table/tbody/tr[1]/td',
+        ).text
+        driver.close()
+        config = {"url": f"{ngrok_url}/github_webhook", "content_type": "json"}
+        gapi = Github(login_or_token=github_token)
+        repo = gapi.get_repo(f"{github_owner}/{github_repository}")
+        for _hook in repo.get_hooks():
+            if "ngrok.io" in _hook.config["url"]:
+                print(f"Deleting existing webhook: {_hook.config['url']}")
+                _hook.delete()
 
     print(
-        f"Creating webhook: {ngrok_url}/github_webhook for "
+        f"Creating webhook: {ngrok_url or webhook_ip}/github_webhook for "
         f"{github_owner}/{github_repository} "
         f"with events: {events}"
     )
